postnotes/requirements-vendor.txt
Django==1.9
google-auth==0.5.0
djangorestframework==3.3.2
python-jose==1.3.2
